Update: 파이어베이스 연결

+++ src/pages/Detail.jsx
@@ -8,10 +8,23 @@ import { deletePost } from "../redux/posts";
 export default function Detail() {
   const { id } = useParams();
   const navigate = useNavigate();
+
   const posts = useSelector((state) => state.posts);
+  const user = useSelector((state) => state.user);
+
   const dispatch = useDispatch();
+
   const post = posts.find((post) => post.id === id);
 
+  const isLoggedIn = () => {
+    if (!user.email) return false;
+    return true;
+  };
+
+  const isSameUser = (author) => {
+    if (user.email !== author) return false;
+    return true;
+  };
   return (
     <>
       <Header />
@@ -44,6 +57,12 @@ export default function Detail() {
         >
           <button
             onClick={() => {
+              if (!isLoggedIn()) {
+                return alert("로그인 후 사용할 수 있습니다.");
+              }
+              if (!isSameUser(post.author)) {
+                return alert("작성자가 일치하지 않습니다.");
+              }
               navigate("/edit", {
                 state: {
                   post,
@@ -64,6 +83,12 @@ export default function Detail() {
           </button>
           <button
             onClick={() => {
+              if (!isLoggedIn()) {
+                return alert("로그인 후 사용할 수 있습니다.");
+              }
+              if (!isSameUser(post.author)) {
+                return alert("작성자가 일치하지 않습니다.");
+              }
               const result = window.confirm("정말로 삭제하시겠습니까?");
               if (result) {
                 dispatch(deletePost(post.id));
